% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_pipeline.r
\name{limma_pipeline}
\alias{limma_pipeline}
\title{limma pipeline function}
\usage{
limma_pipeline(
  expression_matrix,
  group_labels,
  group_contrasts,
  significance_threshold = 0.05,
  reference_group_labels = NULL,
  removeIntercept = FALSE,
  include_unadj_pval = FALSE,
  filter = FALSE,
  normalize = "none",
  limma_trend = FALSE,
  limma_voom = FALSE,
  limma_voom_weight_samples = FALSE,
  output_dir = "./"
)
}
\arguments{
\item{expression_matrix}{A matrix with features as rows and samples as columns}

\item{group_labels}{TODO}

\item{group_contrasts}{TODO}

\item{significance_threshold}{TODO}

\item{reference_group_labels}{TODO}

\item{removeIntercept}{TODO}

\item{include_unadj_pval}{TODO}

\item{filter}{TODO}

\item{normalize}{TODO}

\item{limma_trend}{TODO}

\item{limma_voom}{TODO}

\item{limma_voom_weight_samples}{TODO}

\item{output_dir}{TODO}
}
\description{
limma pipeline function
}
