% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bseq_obj.R
\docType{class}
\name{LRObj-class}
\alias{LRObj-class}
\alias{BObj}
\title{This is a object cointains all the data used and generated by the bulkRNASeq
pipeline}
\description{
This is a object cointains all the data used and generated by the bulkRNASeq
pipeline
}
\section{Slots}{

\describe{
\item{\code{data}}{contains the tables: raw data table and normalized table}

\item{\code{pca}}{contains pca results}

\item{\code{dgenes}}{DE genes results}

\item{\code{parameters}}{contains regarded parameters}

\item{\code{design}}{contains the design matrix}
}}

